rename file and add index.ts
